+++ ptolemy/domains/wireless/demo/SmallWorld/SmallWorldRouter.java
@@ -452,7 +452,7 @@ public class SmallWorldRouter extends TypedAtomicActor {
 
         long seedValue = ((LongToken) (seed.getToken())).longValue();
 
-        if (seedValue != (long) 0) {
+        if (seedValue != 0) {
             _random.setSeed(seedValue);
         } else {
             _random.setSeed(System.currentTimeMillis() + hashCode());
