+++ ptolemy/domains/wireless/lib/network/mac/Backoff.java
@@ -258,7 +258,7 @@ public class Backoff extends MACActorBase {
 
         long sd = ((LongToken) (seed.getToken())).longValue();
 
-        if (sd != (long) 0) {
+        if (sd != 0) {
             _random.setSeed(sd);
         } else {
             _random.setSeed(System.currentTimeMillis() + hashCode());
